app files moved to folder "server" + small refactoring.
