+++ README.md
@@ -1,1 +1,6 @@
 FileSystemWebApp
+
+backend entry point: ./server/server.js
+frontend entry point: ./frontend/js/app.js
+
+root path: ./server/server.js - row 13.
