+++ server/models/viewModels/contentItemViewModel.js
